Remove explicit casting from int to unsigned
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk@411 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/TrussUnit/TrussUnitToolBar.cpp
@@ -169,7 +169,7 @@ void TrussUnitToolBar::moveToolBar ()
     {
         pixNumb -= pixHideNumb;
         pos.setY ( pos.y() - pixHideNumb );
-        if ( pixNumb == (uint)hideButton->getHeight() )
+        if ( pixNumb == hideButton->getHeight() )
             enabled = true;
     }
     setPosition ( pos );
